requirements.txt
keras==2.10.0
matplotlib==3.6.2
numpy==1.22.4
pandas==1.4.2
pandas_datareader==0.10.0
Pillow==9.3.0
plotly==5.11.0
prophet==1.1.1
requests==2.27.1
scikit_learn==1.1.3
streamlit==1.9.0
streamlit_lottie==0.0.3
tensorflow==2.10.0
